rename(dashboard): Errors section -> Exceptions
Section name 'Errors' becomes 'Exceptions', route /errors becomes
/exceptions, file errors.rs becomes exceptions.rs. Metric names
(blockio_errors, etc.) are unchanged — those are accurate at the
kernel level and only the user-facing section copy was rewording.
