+++ crates/dashboard/src/dashboard/exceptions.rs
